[Edit] enum 코드 Constans.swift 파일로 분리
