+++ JUDA_iOS/JUDA/View/LogIn/TermsAndVerificationView.swift
@@ -7,12 +7,6 @@
 
 import SwiftUI
 
-// MARK: - 개인 정보 활용 동의 체크 or 프로필 설정 화면 enum
-enum TermsOrVerification {
-    case TermsOfService
-    case ProfileSetting
-}
-
 // MARK: - 신규 회원: 개인 정보 활용 동의 체크 or 프로필 설정 화면
 struct TermsAndVerificationView: View {
     @State private var viewType: TermsOrVerification = .TermsOfService
