+++ JUDA_iOS/JUDA/View/Mypage/MyPageView.swift
@@ -7,6 +7,7 @@
 
 import SwiftUI
 
+// MARK: - My Page View
 struct MyPageView: View {
     @StateObject private var navigationRouter = NavigationRouter()
     @EnvironmentObject private var authViewModel: AuthViewModel
@@ -45,14 +46,12 @@ struct MyPageView: View {
                                           usedTo: usedTo)
                 case .Record(let recordType):
                     RecordView(recordType: recordType)
-                //
                 case .DrinkDetail(let drink):
                     DrinkDetailView(drink: drink)
                         .modifier(TabBarHidden())
                 case .DrinkDetailWithUsedTo(let drink, let usedTo):
                     DrinkDetailView(drink: drink, usedTo: usedTo)
                         .modifier(TabBarHidden())
-                //
                 case .PostDetail(let postUserType,
                                  let post,
                                  let usedTo):
